refactor: move markChatMessagesAsRead method to messagingService and clean up MessageRepository
