+++ chat/src/main/kotlin/net/thechance/chat/entity/GroupChat.kt
@@ -1,6 +1,12 @@
 package net.thechance.chat.entity
 
-import jakarta.persistence.*
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.ForeignKey
+import jakarta.persistence.Id
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.OneToOne
+import jakarta.persistence.Table
 
 @Entity
 @Table(name = "group_chats", schema = "chat")
